chore: ensure statsig init before sagas init
