+++ src/analytics/ValoraAnalytics.ts
@@ -15,14 +15,10 @@ import {
   isE2EEnv,
   SEGMENT_API_KEY,
   STATSIG_ENV,
-  E2E_TEST_STATSIG_ID,
-  STATSIG_API_KEY,
 } from 'src/config'
 import { store } from 'src/redux/store'
 import Logger from 'src/utils/Logger'
 import { isPresent } from 'src/utils/typescript'
-import { getDefaultStatsigUser } from 'src/statsig'
-import { Statsig } from 'statsig-react-native'
 
 const TAG = 'ValoraAnalytics'
 
@@ -128,21 +124,6 @@ class ValoraAnalytics {
       Logger.info(TAG, 'Segment Analytics Integration initialized!')
     } catch (error) {
       Logger.error(TAG, `Segment setup error: ${error.message}\n`, error)
-    }
-
-    try {
-      const statsigUser = getDefaultStatsigUser()
-      // getAnonymousId causes the e2e tests to fail
-      const overrideStableID = isE2EEnv ? E2E_TEST_STATSIG_ID : await Analytics.getAnonymousId()
-      Logger.debug(TAG, 'Statsig stable ID', overrideStableID)
-      await Statsig.initialize(STATSIG_API_KEY, statsigUser, {
-        // StableID should match Segment anonymousId
-        overrideStableID,
-        environment: STATSIG_ENV,
-        localMode: isE2EEnv,
-      })
-    } catch (error) {
-      Logger.warn(TAG, `Statsig setup error`, error)
     }
   }
